python-ai-service/requirements.txt
Flask
 opencv-python
 mediapipe 
numpy
 Pillow
 requests 
gunicorn

